Add transfer feature

+++ src/main/kotlin/com/shaka/Bank/accout/dto/CreateAccountRequest.kt
@@ -1,9 +1,10 @@
 package com.shaka.Bank.accout.dto
 
 import jakarta.validation.constraints.NotNull
+import java.math.BigDecimal
 
 data class CreateAccountRequest (
     @field:NotNull(message = "User id must be provided")
     val userId: Long,
-    val initialAmount: Double?
+    val initialAmount: BigDecimal?
 )
