+++ src/main/kotlin/com/shaka/Bank/accout/dto/CreateAccountResponse.kt
@@ -1,6 +1,8 @@
 package com.shaka.Bank.accout.dto
 
+import java.math.BigDecimal
+
 data class CreateAccountResponse(
     val accountId: Long,
-    val currentAmount: Double,
+    val currentAmount: BigDecimal,
 )
