+++ src/main/kotlin/com/shaka/Bank/accout/dto/GetAccountBalanceResponse.kt
@@ -1,5 +1,7 @@
 package com.shaka.Bank.accout.dto
 
+import java.math.BigDecimal
+
 data class GetAccountBalanceResponse (
-    val amount: Double,
+    val amount: BigDecimal,
 )
