+++ src/main/kotlin/com/shaka/Bank/transaction/DepositTransaction.kt
@@ -1,12 +1,13 @@
 package com.shaka.Bank.transaction
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class DepositTransaction(
     override val id: Long,
     override val sourceAccountId: Long,
-    override val transactionAmount: Double,
-    override val amountAfterTransaction: Double,
+    override val transactionAmount: BigDecimal,
+    override val amountAfterTransaction: BigDecimal,
     override val timestamp: LocalDateTime,
 ): Transaction {
     override fun isAccountInvolvedInTransaction(accountId: Long): Boolean {
