+++ src/main/kotlin/com/shaka/Bank/transaction/Transaction.kt
@@ -1,12 +1,13 @@
 package com.shaka.Bank.transaction
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 interface Transaction {
     val id: Long
     val sourceAccountId: Long
-    val transactionAmount: Double
-    val amountAfterTransaction: Double
+    val transactionAmount: BigDecimal
+    val amountAfterTransaction: BigDecimal
     val timestamp: LocalDateTime
 
     fun isAccountInvolvedInTransaction(accountId: Long): Boolean
