+++ src/main/kotlin/com/shaka/Bank/transaction/TransferTransaction.kt
@@ -1,13 +1,14 @@
 package com.shaka.Bank.transaction
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class TransferTransaction (
     override val id: Long,
     override val sourceAccountId: Long,
     val destinationAccountId: Long,
-    override val transactionAmount: Double,
-    override val amountAfterTransaction: Double,
+    override val transactionAmount: BigDecimal,
+    override val amountAfterTransaction: BigDecimal,
     override val timestamp: LocalDateTime,
 ): Transaction {
     override fun isAccountInvolvedInTransaction(accountId: Long): Boolean {
