+++ src/main/kotlin/com/shaka/Bank/transaction/dto/DepositRequest.kt
@@ -1,10 +1,11 @@
 package com.shaka.Bank.transaction.dto
 
 import jakarta.validation.constraints.NotNull
+import java.math.BigDecimal
 
 data class DepositRequest (
     @field:NotNull
     val accountId: Long,
     @field:NotNull
-    val amountToDeposit: Double,
+    val amountToDeposit: BigDecimal,
 )
