+++ src/main/kotlin/com/shaka/Bank/transaction/dto/TransactionResponse.kt
@@ -1,5 +1,7 @@
 package com.shaka.Bank.transaction.dto
 
+import java.math.BigDecimal
+
 data class TransactionResponse (
-    val newAccountAmount: Double
+    val newAccountAmount: BigDecimal
 )
